Implement a new select command that should accept a list of fields to display and search criteria.
The search criteria can be compound using "and" or "or" operators. The minimum number of criteria the command should support is two. It is desirable for a command to be able to combine any number of criteria. To simplify coding, only one type of operator is allowed; support for "and" and "or" working together is optional.

The style of record output is tabular. Numbers should be right-aligned, rows should be left-aligned. The size of the table and columns should change automatically depending on the number of characters in the record field.